Add remote debugging via DAP
This commit adds the minimal code needed to have VS Code attach to a
running debug adapter over TCP. The initial approach is to have the Ant
tasks start the server and wait for VS Code to attach to them before
starting the Epsilon script to be debugged.

This only requires using a DebugAdapterDescriptorFactory that looks at
the "port" in the launch.json entry being executed, and passes that on
to the standard DebugAdapterServer class. All the debugging complexity
will be hidden behind the DAP server (currently being prototyped in a
new 'debug-adapter' branch).